requirements.txt
pycryptodome
pymongo
cryptography
pyotp
qrcode[pil]
tkinter